Testing method requires
1. executing the eager-mode model with int64 dtype,
2. executing the ET-VK model with int32 dtype, and
3. narrowing eager-mode result to int32 dtype before comparing.

Differential Revision: [D57649649](https://our.internmc.facebook.com/intern/diff/D57649649/)

[ghstack-poisoned]

+++ .ci/scripts/setup-macos.sh
@@ -28,7 +28,6 @@ install_buck() {
   fi
 
   pushd .ci/docker
-
   # TODO(huydo): This is a one-off copy of buck2 2024-05-15 to unblock Jon and
   # re-enable ShipIt. It’s not ideal that upgrading buck2 will require a manual
   # update the cached binary on S3 bucket too. Let me figure out if there is a
@@ -36,7 +35,11 @@ install_buck() {
   # buck2 only when it’s needed. AFAIK, the complicated part was that buck2
   # --version doesn't say anything w.r.t its release version, i.e. 2024-05-15.
   # See D53878006 for more details.
-  BUCK2=buck2-aarch64-apple-darwin.zst
+  #
+  # If you need to upgrade buck2 version on S3, please reach out to Dev Infra
+  # team for help.
+  BUCK2_VERSION=$(cat ci_commit_pins/buck2.txt)
+  BUCK2=buck2-aarch64-apple-darwin-${BUCK2_VERSION}.zst
   curl -s "https://ossci-macos.s3.amazonaws.com/${BUCK2}" -o "${BUCK2}"
 
   zstd -d "${BUCK2}" -o buck2
@@ -74,16 +77,18 @@ install_sccache() {
 
   export PATH="${SCCACHE_PATH}:${PATH}"
 
-  # Create temp directory for sccache shims
-  TMP_DIR=$(mktemp -d)
-  trap 'rm -rfv ${TMP_DIR}' EXIT
+  # Create temp directory for sccache shims if TMP_DIR doesn't exist
+  if [ -z "${TMP_DIR:-}" ]; then
+    TMP_DIR=$(mktemp -d)
+    trap 'rm -rfv ${TMP_DIR}' EXIT
+    export PATH="${TMP_DIR}:$PATH"
+  fi
 
   write_sccache_stub "${TMP_DIR}/c++"
   write_sccache_stub "${TMP_DIR}/cc"
   write_sccache_stub "${TMP_DIR}/clang++"
   write_sccache_stub "${TMP_DIR}/clang"
 
-  export PATH="${TMP_DIR}:$PATH"
   sccache --zero-stats || true
 }
